Fix dynamic JIT enablement
iOS 17 and above no longer work with the `PT_TRACE_ME` hack, but can
instead enable JIT via a debugging check. However, in our code, we were
unconditionally trying to `ptrace` regardless of what our debugger check
returned.

This works under the XCode debugger, of course, as `ptrace` runs there,
but does not work under any on-device VPN-based JIT enablers (such as
StikDebug). In order to fix this, we made checking for `PT_TRACE_ME`
conditional on _not_ having an attached debugger, which allows the app
to run with JIT if debugged in XCode, on mobile via something like
StikDebug, and without JIT if run on its own.

This is the desired behavior.